requirement.txt
streamlit
tensorflow
opencv-python
scikit-image
numpy
Pillow
